WIP: [docs] Migrate quickstart examples
