+++ apps/docs/src/content/docs/advanced/custom-validation.mdx
@@ -1,7 +1,7 @@
 ---
 title: Custom validation
 sidebar:
-  order: 2
+  order: 1
 ---
 
 ## Errors transformation and custom validation
