+++ apps/docs/src/content/docs/advanced/custom-widgets.mdx
@@ -1,7 +1,7 @@
 ---
 title: Custom widgets
 sidebar:
-  order: 3
+  order: 2
 ---
 
 ## Example
