+++ apps/docs/src/content/docs/advanced/state-transformation.mdx
@@ -1,7 +1,7 @@
 ---
 title: State transformation
 sidebar:
-  order: 4
+  order: 3
 ---
 
 In some cases it may be necessary to transform the form state before it is passed to the validator.
