+++ apps/docs/src/content/docs/guides/_ui-schema.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import type { Schema, UiSchemaRoot } from "@sjsf/form";
+  import { SimpleForm, type Schema, type UiSchemaRoot } from "@sjsf/form";
 
-  import CustomForm from "@/components/custom-form.svelte";
+  import { useCustomForm } from "@/components/custom-form.ts";
 
   const schema: Schema = {
     title: "Schema title",
@@ -17,6 +17,12 @@
       },
     },
   };
+
+  const form = useCustomForm({
+    schema,
+    uiSchema,
+    onSubmit: console.log,
+  });
 </script>
 
-<CustomForm {schema} {uiSchema} onSubmit={console.log} />
+<SimpleForm {form} style="display: flex; flex-direction: column; gap: 1rem" />
