--- apps/docs/src/content/docs/guides/custom-form.mdx
@@ -1,14 +1,0 @@
----
-title: Custom form
-sidebar:
-  order: 4
----
-
-import { Code } from '@astrojs/starlight/components';
-
-import customFormCode from '@/components/custom-form.svelte?raw';
-
-Even with a simple setup, resulting code is very verbose.
-Therefore, after choosing a theme, validator and translation, it is convenient to create a custom component.
-
-<Code code={customFormCode} lang="svelte" title="custom-form.svelte" />
