+++ apps/docs/src/content/docs/guides/quickstart.mdx
@@ -8,10 +8,12 @@ import { Code, Card, LinkCard } from '@astrojs/starlight/components';
 
 
 import MinimalSetup from './_minimal-setup.astro';
-import UiSchema from './_ui-schema.astro';
-import FormData from './_form-data.astro';
 import minimalSetupCode from './_minimal-setup.svelte?raw';
+
+import UiSchema from './_ui-schema.astro';
 import uiSchemaCode from './_ui-schema.svelte?raw';
+
+import FormData from './_form-data.astro';
 import formDataCode from './_form-data.svelte?raw';
 
 ## Minimal setup
